Add files via upload
Updated the binary in order to address an issue some users would face in which the Windows Defender Removal process would terminate almost immediately.  
Complete support and testing on ALL windows 10 and 11 versions, and made sure the tool can be ran by simply right clicking and running it as administrator, or via the cmd line/PowerShell if the user would like more verbose output.  
Just remember to make sure you are running this tool as an administrator or it will not work.  Make sure you also wait until the message appears on screen letting you know it has completed and it is safe to reboot your PC.